Update README.md
Replacing "Agile Software Development Methodology" by "Agile Software Development Mindset".

Agile has been my expertise field for the last 7 years. Based on the Agile community understanding of Agile, I would like to propose to replace Methodology by Mindset.

Agile is a way to deliver value, guided by the principles and values of the Agile Manifesto. It is not a strict methodology with a fixed vocabulary or a step-by-step process,

Not even Scrum defines itself as a methodology but a framework.

Examples of methodologies, IMHO, could be SAFe or PRINCE2 Agile, but no the Agile concept itself.

Thank you for your consideration in advance!
BR,
Marcos.

Signed-off-by: MarcosMG86 <40298360+MarcosMG86@users.noreply.github.com>